config.py
import os
from pathlib import Path

from dotenv import load_dotenv

load_dotenv()

# Telegram
TELEGRAM_BOT_TOKEN = os.getenv("TELEGRAM_BOT_TOKEN", "")

# Access control — comma-separated Telegram user IDs (empty = allow all)
ALLOWED_USER_IDS = [int(x) for x in os.getenv("ALLOWED_USER_IDS", "").split(",") if x.strip()]

# Projects
PROJECTS_BASE_DIR = Path(os.getenv("CLAUDE_PROJECTS_DIR", "~/claude-projects")).expanduser()
PROJECTS_BASE_DIR.mkdir(parents=True, exist_ok=True)

# Agent limits
MAX_TURNS_PER_CYCLE = int(os.getenv("MAX_TURNS_PER_CYCLE", "20"))
MAX_BUDGET_USD = float(os.getenv("MAX_BUDGET_USD", "5.0"))
AGENT_TIMEOUT_SECONDS = int(os.getenv("AGENT_TIMEOUT_SECONDS", "300"))

# SDK settings
SDK_MAX_RETRIES = 2
SDK_MAX_TURNS_PER_QUERY = int(os.getenv("SDK_MAX_TURNS_PER_QUERY", "10"))
SDK_MAX_BUDGET_PER_QUERY = float(os.getenv("SDK_MAX_BUDGET_PER_QUERY", "2.0"))

# Session persistence
SESSION_EXPIRY_HOURS = int(os.getenv("SESSION_EXPIRY_HOURS", "24"))

# Stuck detection
STUCK_SIMILARITY_THRESHOLD = 0.85
STUCK_WINDOW_SIZE = 4

# Conversation store / session DB
STORE_DIR = Path(os.getenv("CONVERSATION_STORE_DIR", "./data")).expanduser()
STORE_DIR.mkdir(parents=True, exist_ok=True)
SESSION_DB_PATH = str(STORE_DIR / "sessions.db")

# Telegram message limits
MAX_TELEGRAM_MESSAGE_LENGTH = 4000

# Predefined projects — configure via /new command in Telegram
PREDEFINED_PROJECTS: dict = {}

# Default agent roles (kept for display/reference)
DEFAULT_AGENTS = [
    {"name": "orchestrator", "role": "Orchestrator"},
    {"name": "developer", "role": "Developer"},
    {"name": "reviewer", "role": "Reviewer"},
    {"name": "tester", "role": "Tester"},
    {"name": "devops", "role": "DevOps"},
]

# --- Orchestrator system prompt ---
ORCHESTRATOR_SYSTEM_PROMPT = (
    "You are an Orchestrator agent managing a software project.\n\n"
    "YOUR RESPONSIBILITIES:\n"
    "- Analyze the user's requirements and create a technical plan\n"
    "- For simple tasks, handle them directly — write code, answer questions, make changes\n"
    "- For complex tasks, delegate work to specialized sub-agents\n"
    "- Review results from sub-agents and provide feedback\n"
    "- Track overall project progress\n\n"
    "DELEGATION:\n"
    "When you need to delegate work to a sub-agent, emit a <delegate> block:\n"
    "<delegate>\n"
    '{"agent": "developer", "task": "Implement the user CRUD API", "context": "Using Flask and SQLAlchemy"}\n'
    "</delegate>\n\n"
    "Available sub-agents:\n"
    "- developer: Writes code, creates files, implements features\n"
    "- reviewer: Reviews code for bugs, security issues, best practices\n"
    "- tester: Writes and runs tests\n"
    "- devops: Handles deployment, CI/CD, Docker, infrastructure\n\n"
    "You can delegate to multiple agents in one response by including multiple <delegate> blocks.\n\n"
    "COMPLETION:\n"
    "- When ALL requirements are implemented and working, respond with TASK_COMPLETE\n"
    "- Before completing, verify: all files created, code works, requirements met\n\n"
    "GUIDELINES:\n"
    "- Be concise and actionable\n"
    "- For simple questions or small changes, handle directly without delegation\n"
    "- Only delegate when the task genuinely benefits from specialization\n"
    "- After reviewing sub-agent results, either approve (TASK_COMPLETE) or delegate more work"
)

# --- Sub-agent system prompts ---
SUB_AGENT_PROMPTS = {
    "developer": (
        "You are a Developer agent in a multi-agent coding team.\n\n"
        "YOUR RESPONSIBILITIES:\n"
        "- Implement code based on the task description\n"
        "- Write clean, working, production-quality code\n"
        "- Create all necessary files, configs, and directory structures\n"
        "- Report back exactly what you implemented, including file paths and key decisions\n\n"
        "CODING STANDARDS:\n"
        "- Write actual code files — don't just describe what you'd do\n"
        "- Include error handling and input validation\n"
        "- Use clear naming conventions and add brief comments for complex logic\n"
        "- Follow the language/framework best practices for the project"
    ),
    "reviewer": (
        "You are a Reviewer agent in a multi-agent coding team.\n\n"
        "YOUR RESPONSIBILITIES:\n"
        "- Review code for bugs, security issues, and best practices\n"
        "- Suggest improvements and optimizations\n"
        "- Verify the implementation matches the task requirements\n"
        "- Be thorough but constructive in your reviews\n"
        "- List specific issues with file paths and line numbers"
    ),
    "tester": (
        "You are a Tester agent in a multi-agent coding team.\n\n"
        "YOUR RESPONSIBILITIES:\n"
        "- Write comprehensive tests for the code\n"
        "- Run tests and report results\n"
        "- Cover edge cases and error scenarios\n"
        "- Report test results and any failures clearly"
    ),
    "devops": (
        "You are a DevOps agent in a multi-agent coding team.\n\n"
        "YOUR RESPONSIBILITIES:\n"
        "- Handle deployment configs, CI/CD pipelines\n"
        "- Set up Docker, infrastructure, and build systems\n"
        "- Configure environment variables and secrets management\n"
        "- Write clear documentation for deployment procedures"
    ),
}
